tests/wpt/tests/css/css-conditional/container-queries/scroll-state/scroll-direction-programmatic-relative-scrolls.html

<!DOCTYPE html>
<title>@container: scroll-state(direction) changed after scroll</title>
<link rel="help" href="https://drafts.csswg.org/css-conditional-5/#direction">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/css-conditional/container-queries/support/cq-testcommon.js"></script>
<script src="/css/css-transitions/support/helper.js"></script>
<style>
  #scroller {
    width: 200px;
    height: 200px;
    container-type: scroll-state;
    overflow-x: scroll;
    overflow-y: scroll;
  }
  #filler {
    height: 600px;
    width: 600px;
  }
  #target {
    --top: no;
    --bottom: no;
    --y: no;
    --left: no;
    --right: no;
    --x: no;
    --none: no;
    @container scroll-state(direction: top) {
      --top: yes;
    }
    @container scroll-state(direction: bottom) {
      --bottom: yes;
    }
    @container scroll-state(direction: y) {
      --y: yes;
    }
    @container scroll-state(direction: left) {
      --left: yes;
    }
    @container scroll-state(direction: right) {
      --right: yes;
    }
    @container scroll-state(direction: x) {
      --x: yes;
    }
    @container scroll-state(direction: none) {
      --none: yes;
    }
  }
</style>
<div id="scroller">
  <div id="filler">
    <div id="target"></div>
  </div>
</div>
<script>
  setup(() => assert_implements_scroll_state_container_queries());

  promise_test(async t => {
    await waitForAnimationFrames(2);
    assert_equals(getComputedStyle(target).getPropertyValue("--none"), "yes");

    assert_equals(getComputedStyle(target).getPropertyValue("--bottom"), "no");
    assert_equals(getComputedStyle(target).getPropertyValue("--top"), "no");
    assert_equals(getComputedStyle(target).getPropertyValue("--y"), "no");

    assert_equals(getComputedStyle(target).getPropertyValue("--left"), "no");
    assert_equals(getComputedStyle(target).getPropertyValue("--right"), "no");
    assert_equals(getComputedStyle(target).getPropertyValue("--x"), "no");
  }, "Check that scroll-state(direction) state before scrolling");

  promise_test(async t => {
    scroller.scrollBy(0, -100);
    await waitForAnimationFrames(2);
    assert_equals(getComputedStyle(target).getPropertyValue("--none"), "yes");

    assert_equals(getComputedStyle(target).getPropertyValue("--bottom"), "no");
    assert_equals(getComputedStyle(target).getPropertyValue("--top"), "no");
    assert_equals(getComputedStyle(target).getPropertyValue("--y"), "no");

    assert_equals(getComputedStyle(target).getPropertyValue("--left"), "no");
    assert_equals(getComputedStyle(target).getPropertyValue("--right"), "no");
    assert_equals(getComputedStyle(target).getPropertyValue("--x"), "no");
  }, "Scrolling up when scrollbar is on the top should not change  scroll-state(direction)");

  promise_test(async t => {
    scroller.scrollTop = 200;
    scroller.scrollBy(0, -100, "instant");
    scroller.scrollTop = 0;
    await waitForAnimationFrames(2);
    assert_equals(getComputedStyle(target).getPropertyValue("--none"), "no");

    assert_equals(getComputedStyle(target).getPropertyValue("--bottom"), "no");
    assert_equals(getComputedStyle(target).getPropertyValue("--top"), "yes");
    assert_equals(getComputedStyle(target).getPropertyValue("--y"), "yes");
    assert_equals(getComputedStyle(target).getPropertyValue("--x"), "no");
  }, "Relative scroll followed by absolute scroll should change scroll direction");

  promise_test(async t => {
    scroller.scrollBy(0, 300);
    await waitForAnimationFrames(2);
    assert_equals(getComputedStyle(target).getPropertyValue("--bottom"), "yes");
    assert_equals(getComputedStyle(target).getPropertyValue("--top"), "no");
    assert_equals(getComputedStyle(target).getPropertyValue("--y"), "yes");
    assert_equals(scroller.scrollTop, 300);
  }, "scroll-state(direction) after scrolling bottom");

  promise_test(async t => {
    scroller.scrollBy(0, -200);
    await waitForAnimationFrames(2);
    assert_equals(getComputedStyle(target).getPropertyValue("--top"), "yes");
    assert_equals(getComputedStyle(target).getPropertyValue("--bottom"), "no");
    assert_equals(getComputedStyle(target).getPropertyValue("--y"), "yes");
    assert_equals(scroller.scrollTop, 100);
  }, "scroll-state(direction) after scrolling top");

  promise_test(async t => {
    scroller.scrollBy(300, 0);
    await waitForAnimationFrames(2);
    assert_equals(getComputedStyle(target).getPropertyValue("--left"), "no");
    assert_equals(getComputedStyle(target).getPropertyValue("--right"), "yes");
    assert_equals(getComputedStyle(target).getPropertyValue("--x"), "yes");
    assert_equals(scroller.scrollLeft, 300);
  }, "scroll-state(direction) after scrolling right");

  promise_test(async t => {
    scroller.scrollBy(-200, 0);
    await waitForAnimationFrames(2);
    assert_equals(getComputedStyle(target).getPropertyValue("--left"), "yes");
    assert_equals(getComputedStyle(target).getPropertyValue("--right"), "no");
    assert_equals(getComputedStyle(target).getPropertyValue("--x"), "yes");
    assert_equals(scroller.scrollLeft, 100);
  }, "scroll-state(direction) after scrolling left");

</script>
